Make size fields of BigIntegerField type.

+++ pulp_2to3_migration/app/plugin/rpm/pulp_2to3_models.py
@@ -89,7 +89,7 @@ class Pulp2RpmBase(Pulp2to3Content):
 
     repodata = JSONField(dict)
     is_modular = models.BooleanField(default=False)
-    size = models.PositiveIntegerField()
+    size = models.BigIntegerField()
     filename = models.TextField()
 
     class Meta:
